Updated all descriptions for clusters and elements for RDL typedefs
Apparently, typedefs do not automatically take on the parent's description. Hence, all the documentation work is somewhat wasted for the examples if they are not updated. Using the trick of changing the typedef from regular to strict and back would mess up all of the nice layouts. Hence, I created "Update Example Control Descriptions.vi" to copy all the documentation work to the clusters and elements without modifying the cluster layout.